requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0

# dev
pre-commit
pytest
pytest-cov

# api-specific reqs
-r src/apis/alteryx_gallery/requirements.txt
-r src/apis/metabase/requirements.txt
-r src/apis/tableau/requirements.txt
